+++ Src/Game/Client.cpp
@@ -12,6 +12,25 @@ std::vector<clientinfo> playerinfos;
 extern void render_map();
 extern std::string name;
 
+sc::time_point last_ping = sc::now();
+
+extern bool can_place(std::vector<tower>& towers, tower_t& base_type, double x, double y);
+
+struct client_state {
+    size_t                  cur_round;
+    size_t                  last_round;
+    size_t                  last_route;
+    std::atomic<double>     lives;
+    bool                    double_speed;
+    bool                    paused;
+    bool                    running;
+    difficulty              diff;
+    uint32_t                spawned_projectiles;
+    std::vector<enemy>      enemies;
+    std::vector<projectile> projectiles;
+    std::vector<tower>      towers;
+};
+
 void do_connect() {
     bool connected = false;
     ENetEvent e { };
@@ -19,6 +38,8 @@ void do_connect() {
     packetstream connection_packet;
     connection_packet << N_CONNECT << (uint32_t)name.length() << name;
     send_packet(peer, 0, true, connection_packet);
+
+    enet_peer_ping_interval(peer, ENET_PEER_PING_INTERVAL);
 
     while(!connected) {
         packetstream p;
@@ -114,9 +135,29 @@ void do_connect() {
 void mouse_press_handler(int x, int y) { }
 void mouse_release_handler(int x, int y) { }
 
+void render_enemies() {
+
+}
+
+void render_towers() {
+
+}
+
+void render_projectiles() {
+
+}
+
+void render_ui() {
+
+}
+
 void render_frame() {
     if(current_map) {
         render_map();
+        render_enemies();
+        render_towers();
+        render_projectiles();
+        render_ui();
     } else {
     //    render_menu();
     }
@@ -124,8 +165,40 @@ void render_frame() {
     render_chat();
 }
 
+void handle_packet(packetstream& p) {
+    uint32_t msgtype = NUMMSG;
+    uint32_t size = -1;
+    p >> msgtype
+      >> size;
+    while(p) {
+        switch(msgtype) {
+
+            default: p.read(nullptr, (size_t)size); break;
+        }
+        p >> msgtype
+          >> size;
+    }
+}
+
 void main_loop() {
     if(peer) {
-         
+        if(sc::now() - last_ping > 500_ms) {
+            packetstream p;
+            p << N_PING << 0_u32;
+            send_packet(peer, 0, true, p);
+        }
+        ENetEvent e;
+        while(enet_host_service(client, &e, maxfps ? 1000_u32 / (uint32_t)maxfps : 1_u32) > 0) {
+            if(e.type == ENET_EVENT_TYPE_RECEIVE) {
+                packetstream stream { (const char*)e.packet->data, e.packet->dataLength };
+                enet_packet_destroy(e.packet);
+                handle_packet(stream);
+            }
+            if(e.type == ENET_EVENT_TYPE_DISCONNECT) {
+                enet_peer_reset(peer);
+                peer = nullptr;
+                break;
+            }
+        }
     }
 }